add delete image functionality to context menu

+++ src/app/services/image.service.ts
@@ -4,6 +4,8 @@ import { AngularFireAuth } from 'angularfire2/auth';
 import { AngularFirestore, AngularFirestoreCollection } from 'angularfire2/firestore';
 import 'firebase/storage';
 import { GalleryImage } from '../models/galleryImage.model';
+import { AngularFireStorage } from '@angular/fire/storage';
+import { firestore } from 'firebase';
 
 
 @Injectable({
@@ -12,8 +14,9 @@ import { GalleryImage } from '../models/galleryImage.model';
 export class ImageService {
   private uid: string;
   private currentAlbum: string;
+  private currentImage: string;
 
-  constructor(private afAuth: AngularFireAuth, private db: AngularFirestore) {
+  constructor(private afAuth: AngularFireAuth, private db: AngularFirestore, private storage: AngularFireStorage) {
     this.afAuth.authState.subscribe(auth => {
       if (auth !== undefined && auth !== null) {
         this.uid = auth.uid;
@@ -28,13 +31,14 @@ export class ImageService {
 
   getImage(imageKey: string, albumKey: string) {
     this.currentAlbum = albumKey;
+    this.currentImage = imageKey;
     return this.db.collection(albumKey).doc(imageKey).ref.get()
       .then(function (doc) {
         if (doc.exists) {
-          return doc.data().url;
+          return doc.data();
         }
         else {     
-          return 'undefined';
+          return undefined;
         }
       })
       .catch(function (error) {
@@ -44,8 +48,21 @@ export class ImageService {
 
   deleteImage(imageKey: string, albumKey: string) {
     this.currentAlbum = albumKey;
-    return this.db.collection(albumKey).doc(imageKey).delete()
-      .then(() => console.log("Image deleteed successfully"))
-      .catch(error => console.log("Error getting image:", error));
+    return this.getImage(imageKey, albumKey)
+      .then(image => {
+        if (image) {
+          const url =   image.url;
+          this.db.collection(albumKey).doc(imageKey).delete()
+            .then(() => console.log("Image deleted successfully"))
+            .catch(error => console.log("Error deleting image:", error))
+            .finally(() => this.deleteFile(url));
+        }
+      })
+      .catch(error => console.log("Error getting image for deletion:", error));
+  }
+
+  private deleteFile(url: string) {
+    const storageRef = this.storage.storage.refFromURL(url);
+    storageRef.delete();
   }
 }
